refactor(game_router): 拆 runtime.py 上帝文件为 session_pool/route_lifecycle/postgame 三子模块
路线图第三批第 7 刀。纯搬移零语义改动：
- session_pool.py: _game_sessions 池 + 创建锁 + key 构建/解析 + 会话构建/刷新/关闭
- route_lifecycle.py: 路由态构建/激活 + 心跳存活 + payload updaters + organizer 调度胶水
- postgame.py: 赛后域（context/event 构建、realtime/文本气泡投递、finalize 主链）
- runtime.py 留 endpoints 与 _run_game_chat 主链，全量回导搬走符号保 monkeypatch 语义
- 唯一向上边 _deliver_postgame_text_bubble→_run_game_chat 改函数内经 runtime 模块属性延迟解析
- gr_patch_all 模块元组加三个新子模块；两个测试文件的本地拷贝改为导入共享 helper
