+++ main_routers/game_router/runtime.py
@@ -13,38 +13,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""The game session/route state machine and every endpoint that
-touches it: session creation, chat, route lifecycle, postgame delivery,
-external transcript routing, end flow and cleanup. Mutable state
-(``_game_sessions``, locks, rate windows) is co-located with all its
-consumers on purpose -- do not split state and consumers apart.
+"""Game chat dispatch and every HTTP endpoint of the game router.
+
+Keeps the endpoint layer plus the ``_run_game_chat`` main chain: session
+chat, route lifecycle endpoints, external transcript routing, end flow and
+the periodic cleanup sweep. The mutable domain state now lives with its
+owning submodule -- ``session_pool`` (``_game_sessions`` and the create
+locks) and ``utils/game_route_state`` (``_game_route_states``) -- and is
+shared by reference; ``route_lifecycle`` owns route-state transitions and
+``postgame`` owns postgame delivery and the finalize flow.
 
 Split out of the former monolithic ``main_routers/game_router.py``.
 """
 
 from ._shared import (
-    _DEFAULT_LAST_FULL_DIALOGUE_COUNT,
     _coerce_payload_bool,
-    _coerce_payload_float,
-    _infer_service_source,
     _log_game_debug_material,
     _normalize_short_text,
     _strip_json_fence,
     logger,
     router,
-)
-from .archive import (
-    _archive_game_context_degraded,
-    _archive_last_assistant_line,
-    _archive_last_user_text,
-    _archive_prompt_language,
-    _archive_score_text,
-    _build_game_archive,
-    _build_game_archive_memory_skipped_result,
-    _fallback_game_archive_memory_highlights,
-    _game_archive_memory_skip_reason,
-    _normalize_game_archive_memory_highlights,
-    _submit_game_archive_to_memory,
 )
 from .badminton_scores import (
     _badminton_end_payload_completed_round,
@@ -56,7 +44,6 @@ from .badminton_scores import (
 from .balance import (
     _apply_badminton_anger_pressure_cap,
     _apply_soccer_anger_pressure_cap,
-    _badminton_duel_difficulty_control_prompt,
     _build_badminton_duel_anger_pressure_cap,
     _build_badminton_duel_balance_hint,
     _build_soccer_anger_pressure_cap,
@@ -70,34 +57,15 @@ from .char_info import (
     _resolve_game_prompt_language,
 )
 from .game_context import (
-    _GAME_CONTEXT_FAILURE_FALLBACK_KEEP_COUNT,
     _GAME_CONTEXT_FAILURE_VISIBLE_WINDOW_MAX_COUNT,
-    _GAME_CONTEXT_FINALIZE_WAIT_SECONDS,
-    _GAME_CONTEXT_ORGANIZE_TRIGGER_COUNT,
-    _GAME_CONTEXT_RECENT_KEEP_COUNT,
-    _GAME_CONTEXT_RECENT_WINDOW_MAX_COUNT,
-    _build_game_context_prompt_payload,
-    _dialog_id_index,
-    _dialog_memory_line,
-    _empty_game_context_signals,
-    _format_game_context_for_prompt,
-    _format_ts,
     _game_context_recent_dialogues,
-    _game_context_signals_text,
-    _merge_game_context_signals,
-    _normalize_game_context_organizer_state,
-    _run_game_context_organizer_ai,
 )
 from .memory_policy import (
-    _DEFAULT_BADMINTON_GAME_MEMORY_ENABLED,
-    _DEFAULT_GAME_MEMORY_TAIL_COUNT,
-    _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
     _attach_game_memory_flag_to_event,
     _game_memory_camel_key,
     _game_memory_player_interaction_enabled,
     _game_memory_policy,
     _game_memory_policy_fields,
-    _game_memory_policy_from_payload,
     _game_memory_postgame_context_enabled,
     _normalize_game_memory_tail_count,
     _normalize_game_memory_type,
@@ -109,36 +77,114 @@ from .pregame import (
     _build_soccer_pregame_context,
     _default_badminton_pregame_context,
     _default_soccer_pregame_context,
-    _format_badminton_pregame_context_for_prompt,
-    _format_soccer_pregame_context_for_prompt,
 )
 from .visible_events import _build_game_llm_visible_event, _sanitize_badminton_event, _sanitize_game_visible_line
+
+# Split domain modules (this file's former state-machine body). Runtime code
+# consumes a subset of these names, but EVERY moved name is re-imported on
+# purpose: callers and tests keep addressing them as ``runtime.<name>``
+# (``gr_patch_all`` then patches every submodule holding the binding), which
+# preserves the pre-split monkeypatch semantics. The mutable containers
+# (``_game_sessions``, ``_game_session_create_locks``) stay the same object
+# as in ``session_pool`` -- from-imports bind by reference, never copy.
+from .postgame import (  # noqa: F401
+    _POSTGAME_REALTIME_NUDGE_DELAYS,
+    _POSTGAME_REALTIME_UNORGANIZED_LIMIT,
+    _POSTGAME_REALTIME_UNORGANIZED_MAX_TOKENS,
+    _POSTGAME_SKIP_REASONS,
+    _active_realtime_session,
+    _append_token_limited_lines,
+    _archive_unorganized_dialogues,
+    _build_game_postgame_context_text,
+    _build_game_postgame_event,
+    _build_game_postgame_realtime_nudge_instruction,
+    _build_postgame_context_snapshot,
+    _deliver_game_postgame,
+    _deliver_postgame_text_bubble,
+    _deliver_postgame_to_realtime,
+    _finalize_game_route_state,
+    _finalize_game_route_state_inner,
+    _is_gemini_realtime_session,
+    _normalize_postgame_options,
+    _postgame_context_request_id,
+    _postgame_last_signals,
+    _run_postgame_realtime_nudge_task,
+)
+from .route_lifecycle import (  # noqa: F401
+    _GAME_ROUTE_ACTIVATION_LOG_LIMIT,
+    _GAME_ROUTE_HEARTBEAT_INTERVAL_SECONDS,
+    _GAME_ROUTE_HEARTBEAT_SWEEP_SECONDS,
+    _GAME_ROUTE_HEARTBEAT_TIMEOUT_SECONDS,
+    _GAME_ROUTE_HIDDEN_HEARTBEAT_TIMEOUT_SECONDS,
+    _GAME_ROUTE_OUTPUT_LIMIT,
+    _activate_game_route,
+    _append_game_dialog,
+    _append_game_output,
+    _append_route_activation,
+    _apply_game_context_failure_fallback,
+    _apply_game_context_organizer_failure,
+    _apply_game_context_organizer_success,
+    _build_route_state,
+    _cancel_game_context_organizer_before_disabled_archive,
+    _detect_before_game_external_state,
+    _find_game_route_state_for_session,
+    _game_context_pending_dialogues,
+    _game_context_recent_id_limit,
+    _maybe_schedule_game_context_organizer,
+    _next_game_dialog_id,
+    _push_game_window_state_change,
+    _route_heartbeat_expired,
+    _route_heartbeat_timeout_seconds,
+    _route_liveness_at,
+    _run_game_context_organizer_task,
+    _set_game_context_recent_ids,
+    _settle_game_context_organizer_before_archive,
+    _should_schedule_game_context_organizer,
+    _sync_game_dialog_seq_from_id,
+    _update_game_memory_enabled_from_payload,
+    _update_route_start_state_from_payload,
+    _update_route_visibility_from_payload,
+)
+from .session_pool import (  # noqa: F401
+    _POSTGAME_SESSION_MARKER,
+    _POSTGAME_UUID_TAIL_RE,
+    _SESSION_CLEANUP_SWEEP_SECONDS,
+    _SESSION_TIMEOUT_SECONDS,
+    _build_and_register_game_session,
+    _build_game_prompt,
+    _close_and_remove_session,
+    _game_session_create_locks,
+    _game_session_key,
+    _game_sessions,
+    _get_or_create_session,
+    _get_session_create_lock,
+    _make_postgame_session_id,
+    _parse_game_session_key,
+    _refresh_game_session_instructions,
+    _route_session_id,
+)
 
 import asyncio
 import json
 import re
 import time
-import uuid
 from collections import OrderedDict
 from typing import Any, Dict, Optional
 from fastapi import HTTPException, Request
 from config.prompts.prompts_soccer import (
     get_soccer_quick_lines_prompt,
     get_soccer_quick_lines_user_prompt,
-    get_soccer_system_prompt,
+    get_soccer_system_prompt,  # noqa: F401  (re-export: paired with the badminton one below)
 )
 from config.prompts.prompts_badminton import (
     get_badminton_quick_lines_fallback,
     get_badminton_quick_lines_prompt,
     get_badminton_quick_lines_user_prompt,
-    get_badminton_system_prompt,
+    get_badminton_system_prompt,  # noqa: F401  (re-export: tests address it as ``runtime.get_badminton_system_prompt``)
 )
 from config.prompts.prompts_minigame_route import (
     get_compact_realtime_context_texts,
     get_game_chat_event_user_prompt,
-    get_game_postgame_context_labels,
-    get_game_postgame_event_texts,
-    get_game_postgame_realtime_nudge_labels,
     get_game_recent_history_message_labels,
 )
 from ..shared_state import get_config_manager, get_session_manager
@@ -162,7 +208,6 @@ from utils.game_log import (
     mark_game_session_debug_log_ended as _mark_game_session_debug_log_ended,
     touch_game_session_debug_log as _touch_game_session_debug_log,
 )
-from utils.language_utils import get_global_language
 
 
 _EXTERNAL_VOICE_DEDUP_TTL_SECONDS = 30.0
@@ -177,19 +222,6 @@ _SSML_TAG_PATTERN = re.compile(
     r")(?:\s+[^<>\n]{0,120})?\s*/?>",
     re.IGNORECASE,
 )
-
-
-# ── Session 池 ─────────────────────────────────────────────────────
-# key = f"{lanlan_name}:{game_type}:{session_id}"
-# value = { session: OmniOfflineClient, reply_chunks: list, last_activity: float, lock: asyncio.Lock }
-_game_sessions: Dict[str, dict] = {}
-
-
-# 超时清理：30 分钟无活动自动销毁
-_SESSION_TIMEOUT_SECONDS = 30 * 60
-
-
-_GAME_ROUTE_ACTIVATION_LOG_LIMIT = 32
 
 
 _SOCCER_QUICK_LINE_KEYS = {
@@ -219,130 +251,6 @@ _BADMINTON_CHAT_RATE_WINDOW_SECONDS = 8.0
 
 
 _BADMINTON_CHAT_RATE_MAX = 10
-
-
-async def _push_game_window_state_change(
-    mgr,
-    *,
-    action: str,
-    lanlan_name: str,
-    game_type: str,
-    session_id: str = "",
-) -> None:
-    """Broadcast the 'game window opened/closed' WS event so the chat.html / pet
-    multi-windows can collapse / restore in sync (user-level UX linkage; not
-    involved in any game-route state decisions, purely drives frontend layout).
-
-    Single source of truth: ``game_route_start`` pushes ``opened`` after
-    activation, and ``_finalize_game_route_state_inner`` pushes ``closed`` after
-    flipping the state to inactive. All finalize paths (/route/end / heartbeat
-    sweep / supersede) go through the inner helper, keeping coverage in sync
-    with ``is_game_route_active`` — the frontend never ends up in an orphaned
-    "game already over but the UI is still locked collapsed" state.
-
-    Multi-window forwarding relies on the existing ``WS_PROXY_CHANNELS.RAW_MESSAGE``
-    IPC (the pet main window receives WS → the forwarder relays it to chat.html),
-    the same bus as mini_game_invite_resolved; no new IPC channel needed.
-    """
-    if not mgr or not lanlan_name:
-        return
-    payload: dict[str, Any] = {
-        "type": "game_window_state_change",
-        "action": action,
-        "lanlan_name": lanlan_name,
-        "game_type": game_type,
-    }
-    if session_id:
-        payload["session_id"] = session_id
-    try:
-        ws = getattr(mgr, "websocket", None)
-        if ws is None or not hasattr(ws, "send_json"):
-            return
-        client_state = getattr(ws, "client_state", None)
-        if client_state is not None and client_state != client_state.CONNECTED:
-            return
-        await ws.send_json(payload)
-    except Exception as exc:
-        logger.warning(
-            "game_window_state_change WS push failed (action=%s, game=%s, lanlan=%s): %s",
-            action, game_type, lanlan_name, exc,
-        )
-
-
-_GAME_ROUTE_OUTPUT_LIMIT = 50
-
-
-_GAME_ROUTE_HEARTBEAT_INTERVAL_SECONDS = 2.5
-
-
-_GAME_ROUTE_HEARTBEAT_TIMEOUT_SECONDS = 10.0
-
-
-_GAME_ROUTE_HIDDEN_HEARTBEAT_TIMEOUT_SECONDS = 60.0
-
-
-_GAME_ROUTE_HEARTBEAT_SWEEP_SECONDS = 2.0
-
-
-_SESSION_CLEANUP_SWEEP_SECONDS = 60.0
-
-
-# Per-(lanlan, game_type, session_id) creation lock for ``_get_or_create_session``.
-#
-# B6: without this, two concurrent ``_run_game_chat`` calls for the same
-# key both miss the cache, both build an ``OmniOfflineClient`` and both
-# ``await session.connect(...)``. The second insertion overwrites
-# ``_game_sessions[key]`` so the first ``entry`` is now an orphan: its
-# ``lock`` is held by the first ``_run_game_chat``, but the cache no
-# longer points at it, so nothing will ever ``close()`` that session.
-#
-# Lifecycle (codex P2 follow-up): the create lock for a key is only
-# meaningful while a session for that key may be created or alive. After
-# ``_close_and_remove_session`` evicts the session from
-# ``_game_sessions``, any further ``_get_or_create_session`` call for
-# the same key would build a fresh session anyway — the lock entry just
-# accumulates without protecting anything. So evict the create lock at
-# the same time as the session, otherwise the dict grows unbounded over
-# uptime as session_ids churn.
-_game_session_create_locks: Dict[str, asyncio.Lock] = {}
-
-
-def _get_session_create_lock(key: str) -> asyncio.Lock:
-    """Lazy-init the per-key creation lock; sync helper, never awaits."""
-    lock = _game_session_create_locks.get(key)
-    if lock is None:
-        lock = _game_session_create_locks.setdefault(key, asyncio.Lock())
-    return lock
-
-
-def _build_game_prompt(
-    game_type: str,
-    lanlan_name: str,
-    lanlan_prompt: str,
-    pre_game_context: dict | None = None,
-    game_context: dict | None = None,
-    language: str | None = None,
-    mode: str = "spectator",
-) -> str:
-    """Build the game system prompt."""
-    if game_type == "soccer":
-        prompt = get_soccer_system_prompt(language).format(name=lanlan_name, personality=lanlan_prompt)
-        context_prompt = _format_soccer_pregame_context_for_prompt(pre_game_context, language)
-        in_game_context_prompt = _format_game_context_for_prompt(game_context, language)
-        return f"{prompt}{context_prompt}{in_game_context_prompt}"
-    if _is_badminton_game_type(game_type):
-        prompt = get_badminton_system_prompt(language, mode=mode).format(name=lanlan_name, personality=lanlan_prompt)
-        if _normalize_badminton_mode(mode) == "duel":
-            prompt = f"{prompt}{_badminton_duel_difficulty_control_prompt(language)}"
-        context_prompt = _format_badminton_pregame_context_for_prompt(pre_game_context, language, mode=mode)
-        in_game_context_prompt = _format_game_context_for_prompt(game_context, language)
-        return f"{prompt}{context_prompt}{in_game_context_prompt}"
-    # 未来其他游戏在这里扩展
-    output_language = str(language or get_global_language() or "en")
-    return (
-        f"You are {lanlan_name}. {lanlan_prompt}\n"
-        f"You are playing a game. Generate short in-character lines in {output_language} for each game event."
-    )
 
 
 def _game_dialog_history_user_text(item: dict, labels: dict[str, str]) -> str:
@@ -529,23 +437,6 @@ def _game_route_closed_session_response(
     return result
 
 
-def _detect_before_game_external_state(mgr: Any) -> tuple[str, bool]:
-    """Return (mode, active) for the current ordinary external session."""
-    if not mgr or not getattr(mgr, "is_active", False):
-        return "none", False
-    session = getattr(mgr, "session", None)
-    try:
-        from main_logic.omni_realtime_client import OmniRealtimeClient
-        from main_logic.omni_offline_client import OmniOfflineClient
-    except Exception:
-        return str(getattr(mgr, "input_mode", "") or "none"), True
-    if isinstance(session, OmniRealtimeClient):
-        return "audio", True
-    if isinstance(session, OmniOfflineClient):
-        return "text", True
-    return str(getattr(mgr, "input_mode", "") or "none"), True
-
-
 def _resolve_lanlan_name(raw: Any = None) -> str:
     lanlan_name = str(raw or "").strip()
     if lanlan_name:
@@ -554,1816 +445,6 @@ def _resolve_lanlan_name(raw: Any = None) -> str:
         return str(_get_current_character_info().get("lanlan_name") or "").strip()
     except Exception:
         return ""
-
-
-def _find_game_route_state_for_session(
-    game_type: str,
-    session_id: str,
-    lanlan_name: str | None = None,
-) -> dict | None:
-    for state in _game_route_states.values():
-        if (
-            str(state.get("game_type") or "") == str(game_type or "")
-            and str(state.get("session_id") or "") == str(session_id or "")
-            and (
-                not lanlan_name
-                or str(state.get("lanlan_name") or "") == str(lanlan_name or "")
-            )
-        ):
-            return state
-    return None
-
-
-def _build_route_state(
-    game_type: str,
-    session_id: str,
-    lanlan_name: str,
-    last_full_dialogue_count: int | None = None,
-) -> dict:
-    session_manager = get_session_manager()
-    mgr = session_manager.get(lanlan_name)
-    before_mode, before_active = _detect_before_game_external_state(mgr)
-    try:
-        keep_last = int(last_full_dialogue_count or _DEFAULT_LAST_FULL_DIALOGUE_COUNT)
-    except (TypeError, ValueError):
-        keep_last = _DEFAULT_LAST_FULL_DIALOGUE_COUNT
-    keep_last = max(1, min(keep_last, 50))
-
-    now = time.time()
-    return {
-        "game_type": game_type,
-        "session_id": session_id,
-        "lanlan_name": lanlan_name,
-        "before_game_external_mode": before_mode,
-        "before_game_external_active": before_active,
-        "game_route_active": True,
-        "game_external_voice_route_active": False,
-        "game_external_text_route_active": False,
-        "game_input_mode": "none",
-        "activation_source": "game_event",
-        "external_suspended_by_game": False,
-        "should_resume_external_on_exit": before_mode == "audio" and before_active,
-        "game_input_activation_log": [],
-        "game_dialog_log": [],
-        "game_dialog_seq": 0,
-        "pending_outputs": [],
-        "game_context_summary": "",
-        "game_context_signals": _empty_game_context_signals(),
-        "game_context_recent_ids": [],
-        "game_context_organizer": {
-            "running": False,
-            "degraded": False,
-            "failure_count": 0,
-            "last_organized_id": "",
-            "source": None,
-            "error": "",
-        },
-        "game_last_full_dialogue_count": keep_last,
-        "game_memory_tail_count": _DEFAULT_GAME_MEMORY_TAIL_COUNT,
-        "soccer_game_memory_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "soccer_game_memory_player_interaction_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "soccer_game_memory_event_reply_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "soccer_game_memory_archive_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "soccer_game_memory_postgame_context_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "badminton_game_memory_enabled": _DEFAULT_BADMINTON_GAME_MEMORY_ENABLED,
-        "badminton_game_memory_player_interaction_enabled": _DEFAULT_BADMINTON_GAME_MEMORY_ENABLED,
-        "badminton_game_memory_event_reply_enabled": _DEFAULT_BADMINTON_GAME_MEMORY_ENABLED,
-        "badminton_game_memory_archive_enabled": _DEFAULT_BADMINTON_GAME_MEMORY_ENABLED,
-        "badminton_game_memory_postgame_context_enabled": _DEFAULT_BADMINTON_GAME_MEMORY_ENABLED,
-        "game_memory_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "game_memory_player_interaction_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "game_memory_event_reply_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "game_memory_archive_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "game_memory_postgame_context_enabled": _DEFAULT_SOCCER_GAME_MEMORY_ENABLED,
-        "last_state": {},
-        "finalScore": {},
-        "preGameContext": {},
-        "pre_game_context_source": "",
-        "pre_game_context_error": "",
-        "nekoInitiated": False,
-        "nekoInviteText": "",
-        "game_started": False,
-        "game_started_at": None,
-        "game_started_elapsed_ms": None,
-        "game_exit_started_elapsed_ms": None,
-        "accidental_game_entry_exit": False,
-        "created_at": now,
-        "last_activity": now,
-        "heartbeat_enabled": True,
-        "last_heartbeat_at": now,
-        "heartbeat_interval_seconds": _GAME_ROUTE_HEARTBEAT_INTERVAL_SECONDS,
-        "heartbeat_timeout_seconds": _GAME_ROUTE_HEARTBEAT_TIMEOUT_SECONDS,
-        "hidden_heartbeat_timeout_seconds": _GAME_ROUTE_HIDDEN_HEARTBEAT_TIMEOUT_SECONDS,
-        "page_visible": True,
-        "visibility_state": "visible",
-    }
-
-
-def _activate_game_route(
-    game_type: str,
-    session_id: str,
-    lanlan_name: str,
-    last_full_dialogue_count: int | None = None,
-) -> dict:
-    state = _build_route_state(game_type, session_id, lanlan_name, last_full_dialogue_count)
-    _game_route_states[_route_state_key(lanlan_name, game_type)] = state
-    logger.info(
-        "🎮 游戏路由已激活: game=%s session=%s lanlan=%s before=%s active=%s",
-        game_type,
-        session_id,
-        lanlan_name,
-        state["before_game_external_mode"],
-        state["before_game_external_active"],
-    )
-    return state
-
-
-def _append_route_activation(state: dict, source: str, mode: str, detail: dict | None = None) -> None:
-    state["game_input_mode"] = mode
-    state["activation_source"] = source
-    state["last_activity"] = time.time()
-    if mode == "voice":
-        state["game_external_voice_route_active"] = True
-    elif mode == "text":
-        state["game_external_text_route_active"] = True
-
-    clean_detail = detail or {}
-    log = state.setdefault("game_input_activation_log", [])
-    if not isinstance(log, list):
-        log = []
-        state["game_input_activation_log"] = log
-
-    # Raw realtime audio arrives as a high-frequency chunk stream.  The
-    # activation log records route mode changes, not every chunk.
-    if not clean_detail:
-        for item in reversed(log):
-            if (
-                isinstance(item, dict)
-                and item.get("source") == source
-                and item.get("mode") == mode
-                and not item.get("detail")
-            ):
-                item["ts"] = state["last_activity"]
-                return
-
-    log.append({
-        "source": source,
-        "mode": mode,
-        "detail": clean_detail,
-        "ts": state["last_activity"],
-    })
-    if len(log) > _GAME_ROUTE_ACTIVATION_LOG_LIMIT:
-        del log[:-_GAME_ROUTE_ACTIVATION_LOG_LIMIT]
-
-
-def _next_game_dialog_id(state: dict) -> str:
-    try:
-        seq = int(state.get("game_dialog_seq") or 0)
-    except (TypeError, ValueError):
-        seq = 0
-    seq += 1
-    state["game_dialog_seq"] = seq
-    return f"glog_{seq:04d}"
-
-
-def _sync_game_dialog_seq_from_id(state: dict, dialog_id: str) -> None:
-    match = re.search(r"(\d+)$", str(dialog_id or ""))
-    if not match:
-        return
-    try:
-        seq = int(match.group(1))
-        current = int(state.get("game_dialog_seq") or 0)
-    except (TypeError, ValueError):
-        current = 0
-        seq = 0
-    if seq > current:
-        state["game_dialog_seq"] = seq
-
-
-def _game_context_pending_dialogues(state: dict) -> list[dict]:
-    dialog = [item for item in state.get("game_dialog_log") or [] if isinstance(item, dict)]
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    last_idx = _dialog_id_index(dialog, str(organizer.get("last_organized_id") or ""))
-    return dialog[last_idx + 1:]
-
-
-def _game_context_recent_id_limit(state: dict, pending_count: int) -> int:
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    if (
-        pending_count > _GAME_CONTEXT_RECENT_WINDOW_MAX_COUNT
-        or organizer.get("running")
-        or int(organizer.get("failure_count") or 0) > 0
-    ):
-        return _GAME_CONTEXT_FAILURE_VISIBLE_WINDOW_MAX_COUNT
-    return _GAME_CONTEXT_RECENT_WINDOW_MAX_COUNT
-
-
-def _apply_game_context_failure_fallback(
-    state: dict,
-    pending: list[dict],
-    *,
-    reason: str,
-) -> bool:
-    if len(pending) < _GAME_CONTEXT_FAILURE_VISIBLE_WINDOW_MAX_COUNT:
-        return False
-    keep_count = _GAME_CONTEXT_FAILURE_FALLBACK_KEEP_COUNT
-    discarded = pending[:-keep_count]
-    kept = pending[-keep_count:]
-    if not discarded:
-        return False
-    last_discarded_id = str(discarded[-1].get("id") or "")
-    if not last_discarded_id:
-        return False
-
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    organizer["last_organized_id"] = last_discarded_id
-    organizer["degraded"] = False
-    organizer["error"] = f"fallback_{reason}_after_{len(pending)}_pending_items"
-    state["game_context_organizer"] = organizer
-    state["game_context_recent_ids"] = [
-        str(item.get("id") or "")
-        for item in kept
-        if isinstance(item, dict) and item.get("id")
-    ]
-    logger.warning(
-        "🎮 局内上下文整理失败兜底丢弃: game=%s session=%s reason=%s discarded=%s kept=%s last=%s",
-        state.get("game_type"),
-        state.get("session_id"),
-        reason,
-        len(discarded),
-        len(kept),
-        last_discarded_id,
-    )
-    return True
-
-
-def _set_game_context_recent_ids(state: dict, dialogues: list[dict] | None = None) -> None:
-    source = dialogues if dialogues is not None else _game_context_pending_dialogues(state)
-    if dialogues is None and len(source) >= _GAME_CONTEXT_FAILURE_VISIBLE_WINDOW_MAX_COUNT:
-        if _apply_game_context_failure_fallback(state, source, reason="overflow"):
-            return
-        source = _game_context_pending_dialogues(state)
-    ids = [str(item.get("id") or "") for item in source if isinstance(item, dict) and item.get("id")]
-    limit = _game_context_recent_id_limit(state, len(ids))
-    state["game_context_recent_ids"] = ids[-limit:]
-
-
-def _should_schedule_game_context_organizer(state: dict) -> bool:
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    state["game_context_organizer"] = organizer
-    if state.get("_exit_flow_started") or state.get("game_route_active") is False:
-        return False
-    if organizer.get("running") or organizer.get("degraded"):
-        return False
-    return len(_game_context_pending_dialogues(state)) >= _GAME_CONTEXT_ORGANIZE_TRIGGER_COUNT
-
-
-def _maybe_schedule_game_context_organizer(state: dict) -> None:
-    """Spawn the per-state organizer task at most once at any given time.
-
-    B4: ``running`` is a dict flag the audit flagged as racy. In practice,
-    on CPython this scheduler is invoked from sync code paths only — the
-    enclosing ``_append_game_dialog`` body has no ``await`` so two
-    coroutines on the same event loop cannot interleave inside it. Still,
-    we add a defensive previous-task done-check so an in-flight organizer
-    is never silently overwritten if a future change introduces an
-    ``await`` boundary in the call chain.
-    """
-    if not _should_schedule_game_context_organizer(state):
-        return
-    prev = state.get("_game_context_organizer_task")
-    if prev is not None and hasattr(prev, "done") and not prev.done():
-        return
-    snapshot = [dict(item) for item in _game_context_pending_dialogues(state)]
-    if len(snapshot) < _GAME_CONTEXT_ORGANIZE_TRIGGER_COUNT:
-        return
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    organizer["running"] = True
-    organizer["error"] = ""
-    state["game_context_organizer"] = organizer
-    try:
-        task = asyncio.create_task(_run_game_context_organizer_task(state, snapshot))
-        state["_game_context_organizer_task"] = task
-    except RuntimeError:
-        organizer["running"] = False
-        state["game_context_organizer"] = organizer
-
-
-def _append_game_dialog(state: dict, item: dict) -> None:
-    # B2: once finalize has started archiving, the snapshot of
-    # ``game_dialog_log`` has already been captured. Mutating it after
-    # that point produces entries that never reach the archive — they
-    # silently disappear when ``_game_route_states`` is eventually
-    # popped by the cleanup sweep. Drop late writes instead.
-    if state.get("_exit_flow_started"):
-        return
-    item = dict(item)
-    item.setdefault("ts", time.time())
-    if item.get("id"):
-        _sync_game_dialog_seq_from_id(state, str(item.get("id") or ""))
-    else:
-        item["id"] = _next_game_dialog_id(state)
-    state.setdefault("game_dialog_log", []).append(item)
-    state["last_activity"] = item["ts"]
-    _set_game_context_recent_ids(state)
-    _maybe_schedule_game_context_organizer(state)
-
-
-def _append_game_output(state: dict, output: dict) -> None:
-    # B2: once finalize has started, ``pending_outputs`` will never be
-    # drained again (the route is exiting, the game page won't ``/drain``
-    # any further). Late writes accumulate into oblivion.
-    if state.get("_exit_flow_started"):
-        return
-    pending = state.setdefault("pending_outputs", [])
-    pending.append(output)
-    del pending[:-_GAME_ROUTE_OUTPUT_LIMIT]
-    state["last_activity"] = time.time()
-
-
-def _apply_game_context_organizer_success(state: dict, snapshot: list[dict], result: dict) -> None:
-    organize_dialogues = snapshot[:-_GAME_CONTEXT_RECENT_KEEP_COUNT]
-    if not organize_dialogues:
-        return
-    target_last_id = str(organize_dialogues[-1].get("id") or "")
-    dialog = [item for item in state.get("game_dialog_log") or [] if isinstance(item, dict)]
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    current_last_id = str(organizer.get("last_organized_id") or "")
-    current_idx = _dialog_id_index(dialog, current_last_id)
-    target_idx = _dialog_id_index(dialog, target_last_id)
-    if current_idx > target_idx >= 0:
-        organizer["running"] = False
-        organizer["error"] = "stale_organizer_result_ignored"
-        state["game_context_organizer"] = organizer
-        _set_game_context_recent_ids(state)
-        return
-
-    summary = _normalize_short_text(
-        result.get("rollingSummary") or result.get("rolling_summary") or result.get("summary"),
-        max_chars=900,
-    )
-    if summary:
-        state["game_context_summary"] = summary
-    state["game_context_signals"] = _merge_game_context_signals(
-        state.get("game_context_signals"),
-        result.get("signals") if isinstance(result.get("signals"), dict) else {},
-    )
-    organizer.update({
-        "running": False,
-        "degraded": False,
-        "failure_count": 0,
-        "last_organized_id": target_last_id,
-        "source": result.get("source") if isinstance(result.get("source"), dict) else result.get("source"),
-        "error": "",
-    })
-    state["game_context_organizer"] = organizer
-    _set_game_context_recent_ids(state)
-
-
-def _apply_game_context_organizer_failure(state: dict, snapshot: list[dict], error: Exception) -> None:
-    organize_dialogues = snapshot[:-_GAME_CONTEXT_RECENT_KEEP_COUNT]
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    if organize_dialogues:
-        target_last_id = str(organize_dialogues[-1].get("id") or "")
-        dialog = [item for item in state.get("game_dialog_log") or [] if isinstance(item, dict)]
-        current_last_id = str(organizer.get("last_organized_id") or "")
-        current_idx = _dialog_id_index(dialog, current_last_id)
-        target_idx = _dialog_id_index(dialog, target_last_id)
-        if current_idx > target_idx >= 0:
-            organizer["running"] = False
-            organizer["error"] = organizer.get("error") or "stale_organizer_failure_ignored"
-            state["game_context_organizer"] = organizer
-            _set_game_context_recent_ids(state)
-            return
-    organizer["running"] = False
-    organizer["failure_count"] = int(organizer.get("failure_count") or 0) + 1
-    organizer["error"] = type(error).__name__
-    state["game_context_organizer"] = organizer
-    pending = _game_context_pending_dialogues(state)
-    fallback_reason = f"organizer_failure_{type(error).__name__}"
-    if _apply_game_context_failure_fallback(state, pending, reason=fallback_reason):
-        return
-    _set_game_context_recent_ids(state)
-
-
-async def _run_game_context_organizer_task(state: dict, snapshot: list[dict]) -> None:
-    succeeded = False
-    try:
-        result = await _run_game_context_organizer_ai(state, snapshot)
-        _apply_game_context_organizer_success(state, snapshot, result)
-        succeeded = True
-    except Exception as exc:
-        _apply_game_context_organizer_failure(state, snapshot, exc)
-    finally:
-        organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-        if organizer.get("running"):
-            organizer["running"] = False
-            state["game_context_organizer"] = organizer
-        if succeeded and not organizer.get("degraded"):
-            _maybe_schedule_game_context_organizer(state)
-
-
-async def _settle_game_context_organizer_before_archive(state: dict) -> None:
-    task = state.get("_game_context_organizer_task")
-    if task is None or not hasattr(task, "done"):
-        return
-
-    if task.done():
-        if task.cancelled():
-            organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-            organizer["running"] = False
-            organizer["error"] = organizer.get("error") or "cancelled"
-            state["game_context_organizer"] = organizer
-            return
-        try:
-            await task
-        except Exception as exc:
-            organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-            organizer["running"] = False
-            organizer["error"] = type(exc).__name__
-            state["game_context_organizer"] = organizer
-            logger.warning(
-                "🎮 退出前收敛局内上下文整理失败: game=%s session=%s err=%s",
-                state.get("game_type"),
-                state.get("session_id"),
-                exc,
-            )
-        return
-
-    try:
-        await asyncio.wait_for(task, timeout=_GAME_CONTEXT_FINALIZE_WAIT_SECONDS)
-    except asyncio.TimeoutError:
-        organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-        organizer["running"] = False
-        organizer["error"] = "finalize_timeout"
-        state["game_context_organizer"] = organizer
-        logger.warning(
-            "🎮 退出前等待局内上下文整理超时，使用已有信息归档: game=%s session=%s timeout=%.1fs",
-            state.get("game_type"),
-            state.get("session_id"),
-            _GAME_CONTEXT_FINALIZE_WAIT_SECONDS,
-        )
-    except Exception as exc:
-        organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-        organizer["running"] = False
-        organizer["error"] = type(exc).__name__
-        state["game_context_organizer"] = organizer
-        logger.warning(
-            "🎮 退出前等待局内上下文整理失败，使用已有信息归档: game=%s session=%s err=%s",
-            state.get("game_type"),
-            state.get("session_id"),
-            exc,
-        )
-
-
-async def _cancel_game_context_organizer_before_disabled_archive(state: dict) -> None:
-    task = state.get("_game_context_organizer_task")
-    organizer = _normalize_game_context_organizer_state(state.get("game_context_organizer"))
-    organizer["running"] = False
-    organizer["error"] = "archive_disabled"
-    state["game_context_organizer"] = organizer
-
-    if task is None or not hasattr(task, "done") or task.done():
-        return
-    task.cancel()
-    try:
-        await task
-    except asyncio.CancelledError:
-        pass
-    except Exception as exc:
-        logger.debug(
-            "🎮 关闭游戏记忆后取消局内上下文整理失败: game=%s session=%s err=%s",
-            state.get("game_type"),
-            state.get("session_id"),
-            exc,
-            exc_info=True,
-        )
-
-
-def _route_liveness_at(state: dict) -> float:
-    """Return the timestamp proving the game page heartbeat is still alive."""
-    for key in ("last_heartbeat_at", "created_at"):
-        try:
-            value = float(state.get(key) or 0)
-        except (TypeError, ValueError):
-            continue
-        if value > 0:
-            return value
-    return 0.0
-
-
-def _route_heartbeat_expired(state: dict, now: float) -> bool:
-    return now - _route_liveness_at(state) > _route_heartbeat_timeout_seconds(state)
-
-
-def _route_heartbeat_timeout_seconds(state: dict) -> float:
-    """Use a longer grace window while the browser reports the game tab hidden."""
-    visibility = str(state.get("visibility_state") or "").strip().lower()
-    page_visible = state.get("page_visible")
-    hidden = page_visible is False or visibility in {"hidden", "prerender", "unloaded"}
-    key = "hidden_heartbeat_timeout_seconds" if hidden else "heartbeat_timeout_seconds"
-    fallback = _GAME_ROUTE_HIDDEN_HEARTBEAT_TIMEOUT_SECONDS if hidden else _GAME_ROUTE_HEARTBEAT_TIMEOUT_SECONDS
-    try:
-        return max(1.0, float(state.get(key, fallback) or fallback))
-    except (TypeError, ValueError):
-        return fallback
-
-
-def _update_route_visibility_from_payload(state: dict, data: dict) -> None:
-    visibility = str(data.get("visibilityState") or data.get("visibility_state") or "").strip().lower()
-    if visibility:
-        state["visibility_state"] = visibility[:32]
-
-    page_visible = data.get("pageVisible")
-    if isinstance(page_visible, bool):
-        state["page_visible"] = page_visible
-    elif visibility:
-        state["page_visible"] = visibility == "visible"
-
-
-def _update_game_memory_enabled_from_payload(
-    state: dict,
-    data: dict,
-    game_type: str | None = None,
-) -> None:
-    gt = _normalize_game_memory_type(game_type or state.get("game_type") or "soccer")
-    policy = _game_memory_policy_from_payload(gt, data, current=state)
-    if policy is not None:
-        fields = _game_memory_policy_fields(gt)
-        for field in fields:
-            state[field] = policy[field]
-        state["game_memory_enabled"] = policy[fields[0]]
-        state["gameMemoryEnabled"] = policy[fields[0]]
-        state["game_memory_player_interaction_enabled"] = policy[fields[1]]
-        state["game_memory_event_reply_enabled"] = policy[fields[2]]
-        state["game_memory_archive_enabled"] = policy[fields[3]]
-        state["game_memory_postgame_context_enabled"] = policy[fields[4]]
-
-
-def _update_route_start_state_from_payload(state: dict, data: dict, *, exiting: bool = False) -> bool:
-    """Track whether the user actually clicked the game Start button."""
-    was_started = state.get("game_started") is True
-    started_value = None
-    if "gameStarted" in data:
-        started_value = _coerce_payload_bool(data.get("gameStarted"))
-    elif "game_started" in data:
-        started_value = _coerce_payload_bool(data.get("game_started"))
-
-    elapsed_ms = None
-    for key in ("gameStartedElapsedMs", "game_started_elapsed_ms"):
-        if key in data:
-            elapsed_ms = _coerce_payload_float(data.get(key))
-            break
-    if elapsed_ms is not None:
-        elapsed_ms = max(0.0, elapsed_ms)
-        state["game_started_elapsed_ms"] = elapsed_ms
-        if exiting:
-            state["game_exit_started_elapsed_ms"] = elapsed_ms
-
-    if started_value is True:
-        state["game_started"] = True
-        if not was_started:
-            state["game_started_at"] = time.time() - ((elapsed_ms or 0.0) / 1000.0)
-    elif started_value is False and not was_started:
-        state["game_started"] = False
-
-    accidental = _coerce_payload_bool(data.get("accidentalGameEntry"))
-    if accidental is None:
-        accidental = _coerce_payload_bool(data.get("accidental_game_entry"))
-    if accidental is True:
-        state["accidental_game_entry_exit"] = True
-
-    started_now = not was_started and state.get("game_started") is True
-    if started_now:
-        # 在 game_started 首次 false→true 的边沿统计游玩次数——不在 /route/start
-        # 计数：前端 _prepareGameForStartScreen 会先打开开始屏并调 route/start，
-        # 此时 game_started 仍为 false，用户若从开始屏关闭会被记 accidental_page_entry，
-        # 那种"开了没玩"不应计入。本函数是所有上报 gameStarted 路径的唯一汇聚点，
-        # was_started 守卫保证每局只记一次。game_type 从 state 取（route/start 已写入）。
-        #
-        # 不带 neko_initiated 维度：state["nekoInitiated"] 只来自 route/start payload，
-        # 而邀请被接受后 window.open(game_url) 只透传 lanlan_name/session_id、不回填
-        # nekoInitiated，故邀请局会被误标 false。要修准要么动 nekoInitiated（同时驱动
-        # pregame 语气分析，越界）要么跨三端加 from_invite 管线（无法充分验证 Electron）。
-        # 该维度本非需求项，宁缺毋滥；邀请→游玩转化由 mini_game_invited 与本计数的总量得出。
-        try:
-            from utils.instrument import counter as _instr_counter
-            _instr_counter(
-                "mini_game_played",
-                game_type=str(state.get("game_type") or "")[:24],
-            )
-        except Exception:
-            # 埋点 best-effort，失败不影响游戏状态机
-            pass
-    return started_now
-
-
-_POSTGAME_SKIP_REASONS = {"heartbeat_timeout", "session_cleanup", "cleanup", "manual_return_to_start"}
-
-
-_POSTGAME_REALTIME_NUDGE_DELAYS = (1.5, 5.0, 9.0)
-
-
-_POSTGAME_REALTIME_UNORGANIZED_LIMIT = 12
-
-
-_POSTGAME_REALTIME_UNORGANIZED_MAX_TOKENS = 1500
-
-
-def _normalize_postgame_options(raw: Any, *, reason: str) -> dict:
-    """Normalize one-shot postgame delivery options from the game-end request."""
-    reason_text = str(reason or "").strip().lower()
-    options = {
-        "enabled": reason_text not in _POSTGAME_SKIP_REASONS,
-        "mode": "auto",
-        "trigger_voice": True,
-        "include_last_dialogues": _DEFAULT_LAST_FULL_DIALOGUE_COUNT,
-        "max_chars": 60,
-        "min_idle_secs": 0.0,
-        "force_on_skip_reason": False,
-    }
-    if raw is False:
-        options["enabled"] = False
-    elif isinstance(raw, dict):
-        if "enabled" in raw:
-            options["enabled"] = bool(raw.get("enabled"))
-        mode = str(raw.get("mode") or "").strip().lower()
-        if mode in {"auto", "realtime", "text", "off"}:
-            options["mode"] = mode
-        if options["mode"] == "off":
-            options["enabled"] = False
-        if "triggerVoice" in raw:
-            options["trigger_voice"] = bool(raw.get("triggerVoice"))
-        elif "trigger_voice" in raw:
-            options["trigger_voice"] = bool(raw.get("trigger_voice"))
-        if "forceOnSkipReason" in raw:
-            options["force_on_skip_reason"] = bool(raw.get("forceOnSkipReason"))
-        for source_key, target_key, low, high in (
-            ("includeLastDialogues", "include_last_dialogues", 1, 50),
-            ("maxChars", "max_chars", 20, 160),
-        ):
-            if source_key in raw:
-                try:
-                    options[target_key] = max(low, min(int(raw.get(source_key)), high))
-                except (TypeError, ValueError):
-                    pass
-        if "minIdleSecs" in raw:
-            try:
-                options["min_idle_secs"] = max(0.0, min(float(raw.get("minIdleSecs")), 30.0))
-            except (TypeError, ValueError):
-                pass
-
-    if reason_text in _POSTGAME_SKIP_REASONS and not options["force_on_skip_reason"]:
-        options["enabled"] = False
-    return options
-
-
-def _postgame_last_signals(archive: dict) -> dict:
-    dialogues = archive.get("last_full_dialogues") if isinstance(archive.get("last_full_dialogues"), list) else []
-    signals = {
-        "last_user_text": "",
-        "last_assistant_line": "",
-        "final_mood": "",
-        "final_difficulty": "",
-    }
-    for item in reversed(dialogues):
-        if not isinstance(item, dict):
-            continue
-        if not signals["last_user_text"] and item.get("type") == "user":
-            signals["last_user_text"] = str(item.get("text") or "").strip()
-        if not signals["last_assistant_line"]:
-            signals["last_assistant_line"] = str(item.get("line") or item.get("result_line") or "").strip()
-        control = item.get("control") if isinstance(item.get("control"), dict) else {}
-        if not signals["final_mood"] and control.get("mood"):
-            signals["final_mood"] = str(control.get("mood") or "").strip()
-        if not signals["final_difficulty"] and control.get("difficulty"):
-            signals["final_difficulty"] = str(control.get("difficulty") or "").strip()
-        if all(signals.values()):
-            break
-    return signals
-
-
-def _archive_unorganized_dialogues(archive: dict, *, limit: int = _POSTGAME_REALTIME_UNORGANIZED_LIMIT) -> list[dict]:
-    dialogues = archive.get("full_dialogues") if isinstance(archive.get("full_dialogues"), list) else []
-    dialogues = [item for item in dialogues if isinstance(item, dict)]
-    if not dialogues:
-        last_dialogues = archive.get("last_full_dialogues") if isinstance(archive.get("last_full_dialogues"), list) else []
-        dialogues = [
-            item
-            for item in last_dialogues
-            if isinstance(item, dict)
-        ]
-    organizer = _normalize_game_context_organizer_state(archive.get("game_context_organizer"))
-    last_idx = _dialog_id_index(dialogues, str(organizer.get("last_organized_id") or ""))
-    pending = dialogues[last_idx + 1:] if last_idx >= 0 else dialogues
-    return pending[-max(1, limit):]
-
-
-def _append_token_limited_lines(lines: list[str], header: str, raw_lines: list[str], *, max_tokens: int) -> None:
-    from utils.tokenize import count_tokens, truncate_to_tokens
-
-    if max_tokens <= 0:
-        return
-
-    kept: list[str] = []
-    total_tokens = 0
-    for raw in reversed(raw_lines):
-        line = str(raw or "").strip()
-        if not line:
-            continue
-        line_tokens = count_tokens(line)
-        next_total = total_tokens + line_tokens
-        if next_total > max_tokens:
-            # Even the first (newest) line must respect the budget — a single
-            # pasted/long dialogue entry would otherwise bypass the cap.
-            if not kept:
-                clipped = truncate_to_tokens(line, max_tokens)
-                if clipped:
-                    kept.insert(0, clipped)
-            break
-        kept.insert(0, line)
-        total_tokens = next_total
-    if kept:
-        lines.append(header)
-        lines.extend(kept)
-
-
-def _build_game_postgame_context_text(archive: dict) -> str:
-    """Context for an already-active Realtime session; it should not speak by itself.
-
-    Reuse already-built game archive material only. Do not trigger another LLM
-    pass here; the Realtime session only needs compact postgame continuity.
-    """
-    language = _archive_prompt_language(archive)
-    labels = get_game_postgame_context_labels(language)
-    degraded = _archive_game_context_degraded(archive)
-    score_text = _archive_score_text(archive)
-    highlights = _normalize_game_archive_memory_highlights(archive.get("memory_highlights"))
-    if not any(
-        (
-            highlights["important_records"],
-            highlights["important_game_events"],
-            highlights["state_carryback"],
-            highlights["postgame_tone"],
-            highlights["memory_summary"],
-        )
-    ):
-        highlights = _normalize_game_archive_memory_highlights(_fallback_game_archive_memory_highlights(archive))
-
-    lines = [
-        labels["header"],
-        labels["description"],
-        labels["usage"],
-        labels["game"].format(game_type=archive.get("game_type") or "game"),
-        labels["session"].format(session_id=archive.get("session_id") or "default"),
-        labels["time"].format(start=_format_ts(archive.get("created_at")), end=_format_ts(archive.get("ended_at"))),
-    ]
-    if score_text:
-        lines.append(labels["official_result"].format(score_text=score_text))
-    summary = str(archive.get("summary") or "").strip()
-    if summary:
-        lines.append(labels["summary"].format(summary=summary))
-    lines.append(labels["result_rule"])
-
-    if degraded:
-        lines.append(labels["degraded"])
-    else:
-        if highlights["memory_summary"]:
-            lines.append(labels["memory_summary"].format(value=highlights["memory_summary"]))
-        if highlights["important_records"]:
-            lines.append(labels["important_records"])
-            lines.extend(f"- {item}" for item in highlights["important_records"])
-        if highlights["important_game_events"]:
-            lines.append(labels["important_game_events"])
-            lines.extend(f"- {item}" for item in highlights["important_game_events"])
-        if highlights["state_carryback"]:
-            lines.append(labels["state_carryback"].format(value=highlights["state_carryback"]))
-        if highlights["postgame_tone"]:
-            lines.append(labels["postgame_tone"].format(value=highlights["postgame_tone"]))
-
-        context_summary = _normalize_short_text(archive.get("game_context_summary"), max_chars=900)
-        signals_text = _game_context_signals_text(archive.get("game_context_signals"))
-        if context_summary:
-            lines.append(labels["rolling_summary"].format(summary=context_summary))
-        if signals_text:
-            lines.append(labels["signals"])
-            lines.append(signals_text)
-
-    unorganized_lines = [
-        f"- {_dialog_memory_line(item, language)}"
-        for item in _archive_unorganized_dialogues(archive)
-        if isinstance(item, dict)
-    ]
-    _append_token_limited_lines(
-        lines,
-        labels["unorganized_window"],
-        unorganized_lines,
-        max_tokens=_POSTGAME_REALTIME_UNORGANIZED_MAX_TOKENS,
-    )
-
-    last_user = _archive_last_user_text(archive)
-    last_assistant = _archive_last_assistant_line(archive)
-    if last_user:
-        lines.append(labels["last_user"].format(text=last_user))
-    if last_assistant:
-        lines.append(labels["last_assistant"].format(text=last_assistant))
-
-    lines.append(labels["reply_rule"])
-    return "\n".join(line for line in lines if line is not None)
-
-
-def _build_game_postgame_realtime_nudge_instruction(archive: dict, options: dict) -> str:
-    labels = get_game_postgame_realtime_nudge_labels(_archive_prompt_language(archive))
-    signals = _postgame_last_signals(archive)
-    max_chars = int(options.get("max_chars") or 60)
-    degraded = _archive_game_context_degraded(archive)
-    lines = [
-        labels["header"],
-        labels["ended"],
-        labels["no_ingame"],
-    ]
-    summary = str(archive.get("summary") or "").strip()
-    if summary:
-        lines.append(labels["summary"].format(summary=summary))
-    score_text = _archive_score_text(archive)
-    if score_text:
-        lines.append(labels["score"].format(score_text=score_text))
-        lines.append(labels["score_rule"])
-    if degraded:
-        lines.append(labels["degraded"])
-    if signals["last_user_text"]:
-        lines.append(labels["last_user"].format(text=signals["last_user_text"]))
-    if signals["last_assistant_line"]:
-        lines.append(labels["last_assistant"].format(text=signals["last_assistant_line"]))
-    highlights = _normalize_game_archive_memory_highlights(archive.get("memory_highlights"))
-    if highlights["state_carryback"] and not degraded:
-        lines.append(labels["state_carryback"].format(value=highlights["state_carryback"]))
-    if highlights["postgame_tone"] and not degraded:
-        lines.append(labels["postgame_tone"].format(value=highlights["postgame_tone"]))
-    lines.append(labels["request"].format(max_chars=max_chars))
-    return "\n".join(lines)
-
-
-def _build_game_postgame_event(game_type: str, archive: dict, options: dict) -> dict:
-    language = _archive_prompt_language(archive)
-    texts = get_game_postgame_event_texts(language)
-    dialogues = archive.get("last_full_dialogues") if isinstance(archive.get("last_full_dialogues"), list) else []
-    include_count = int(options.get("include_last_dialogues") or _DEFAULT_LAST_FULL_DIALOGUE_COUNT)
-    formatted_dialogues = [
-        _dialog_memory_line(item, language)
-        for item in dialogues[-include_count:]
-        if isinstance(item, dict)
-    ]
-    signals = _postgame_last_signals(archive)
-    current_state = dict(archive.get("last_state") or {}) if isinstance(archive.get("last_state"), dict) else {}
-    final_score = archive.get("finalScore") if isinstance(archive.get("finalScore"), dict) else {}
-    if final_score:
-        current_state["score"] = dict(final_score)
-    return {
-        "kind": "postgame",
-        "lanlan_name": archive.get("lanlan_name") or "",
-        "label": texts["label"],
-        "gameType": game_type,
-        "summary": archive.get("summary") or "",
-        "scoreText": _archive_score_text(archive),
-        "finalScore": final_score,
-        "lastDialogues": formatted_dialogues,
-        "lastUserText": signals["last_user_text"],
-        "lastAssistantLine": signals["last_assistant_line"],
-        "finalMood": signals["final_mood"],
-        "finalDifficulty": signals["final_difficulty"],
-        "currentState": current_state,
-        "preGameContext": archive.get("preGameContext") if isinstance(archive.get("preGameContext"), dict) else {},
-        "memoryHighlights": _normalize_game_archive_memory_highlights(archive.get("memory_highlights")),
-        "request": texts["request"].format(max_chars=int(options.get("max_chars") or 60)),
-    }
-
-
-def _active_realtime_session(mgr: Any) -> Any | None:
-    if not (mgr and getattr(mgr, "is_active", False)):
-        return None
-    session = getattr(mgr, "session", None)
-    try:
-        from main_logic.omni_realtime_client import OmniRealtimeClient
-    except Exception:
-        return None
-    return session if isinstance(session, OmniRealtimeClient) else None
-
-
-def _is_gemini_realtime_session(session: Any) -> bool:
-    return bool(getattr(session, "_is_gemini", False))
-
-
-async def _run_postgame_realtime_nudge_task(
-    mgr: Any,
-    archive: dict,
-    options: dict,
-    delays: tuple[float, ...],
-    *,
-    expected_session: Any | None = None,
-) -> None:
-    lanlan_name = str(archive.get("lanlan_name") or "")
-    instruction = _build_game_postgame_realtime_nudge_instruction(archive, options)
-    _log_game_debug_material(
-        "postgame_realtime_nudge_instruction",
-        instruction,
-        game_type=str(archive.get("game_type") or ""),
-        session_id=str(archive.get("session_id") or ""),
-        lanlan_name=lanlan_name,
-        source="game_end",
-    )
-    for attempt, delay in enumerate(delays, start=1):
-        try:
-            await asyncio.sleep(delay)
-            active_session = _active_realtime_session(mgr)
-            if not active_session:
-                logger.info(
-                    "🎮 赛后 Realtime 主动搭话跳过: game=%s session=%s lanlan=%s attempt=%d reason=no_active_realtime_session",
-                    archive.get("game_type"),
-                    archive.get("session_id"),
-                    lanlan_name,
-                    attempt,
-                )
-                return
-            if expected_session is not None and active_session is not expected_session:
-                logger.info(
-                    "🎮 赛后 Realtime 主动搭话跳过: game=%s session=%s lanlan=%s attempt=%d reason=realtime_session_changed",
-                    archive.get("game_type"),
-                    archive.get("session_id"),
-                    lanlan_name,
-                    attempt,
-                )
-                return
-
-            trigger = getattr(mgr, "trigger_voice_proactive_nudge", None)
-            if not callable(trigger):
-                logger.info(
-                    "🎮 赛后 Realtime 主动搭话跳过: game=%s session=%s lanlan=%s attempt=%d reason=trigger_unavailable",
-                    archive.get("game_type"),
-                    archive.get("session_id"),
-                    lanlan_name,
-                    attempt,
-                )
-                return
-
-            delivered = bool(await trigger())
-            logger.info(
-                "🎮 赛后 Realtime 主动搭话尝试: game=%s session=%s lanlan=%s attempt=%d delay=%.1fs delivered=%s",
-                archive.get("game_type"),
-                archive.get("session_id"),
-                lanlan_name,
-                attempt,
-                delay,
-                delivered,
-            )
-            if delivered:
-                return
-        except asyncio.CancelledError:
-            raise
-        except Exception as exc:
-            logger.warning(
-                "🎮 赛后 Realtime 主动搭话异常: game=%s session=%s lanlan=%s attempt=%d err=%s",
-                archive.get("game_type"),
-                archive.get("session_id"),
-                lanlan_name,
-                attempt,
-                exc,
-            )
-    logger.info(
-        "🎮 赛后 Realtime 主动搭话放弃: game=%s session=%s lanlan=%s attempts=%d",
-        archive.get("game_type"),
-        archive.get("session_id"),
-        lanlan_name,
-        len(delays),
-    )
-
-
-def _postgame_context_request_id(archive: dict) -> Optional[str]:
-    game_type = str(archive.get("game_type") or "game").strip() or "game"
-    session_id = str(archive.get("session_id") or "default").strip() or "default"
-    ended_at = str(archive.get("ended_at") or "").strip()
-    if not ended_at:
-        return None
-    return f"{game_type}:{session_id}:{ended_at}"
-
-
-async def _deliver_postgame_to_realtime(mgr: Any, archive: dict, options: dict) -> dict:
-    session = _active_realtime_session(mgr)
-    if not session:
-        return {"ok": False, "mode": "realtime", "action": "skip", "reason": "no_active_realtime_session"}
-
-    text = _build_game_postgame_context_text(archive)
-    _log_game_debug_material(
-        "postgame_realtime_context",
-        text,
-        game_type=str(archive.get("game_type") or ""),
-        session_id=str(archive.get("session_id") or ""),
-        lanlan_name=str(archive.get("lanlan_name") or ""),
-        source="game_end",
-    )
-    if _active_realtime_session(mgr) is not session:
-        return {
-            "ok": False,
-            "mode": "realtime",
-            "action": "skip",
-            "reason": "realtime_session_changed",
-        }
-
-    if _is_gemini_realtime_session(session):
-        instruction = _build_game_postgame_realtime_nudge_instruction(archive, options)
-        _log_game_debug_material(
-            "postgame_realtime_nudge_instruction",
-            instruction,
-            game_type=str(archive.get("game_type") or ""),
-            session_id=str(archive.get("session_id") or ""),
-            lanlan_name=str(archive.get("lanlan_name") or ""),
-            source="game_end",
-        )
-        if not options.get("trigger_voice", True):
-            return {
-                "ok": True,
-                "mode": "realtime",
-                "action": "skip",
-                "reason": "gemini_direct_response_disabled",
-                "context_injected": False,
-                "nudge_scheduled": False,
-            }
-        create_response = getattr(session, "create_response", None)
-        if not callable(create_response):
-            return {
-                "ok": False,
-                "mode": "realtime",
-                "action": "skip",
-                "reason": "gemini_create_response_unavailable",
-            }
-        try:
-            await create_response(text + "\n\n" + instruction)
-        except Exception as exc:
-            logger.warning(
-                "🎮 赛后 Gemini Realtime 直接触发失败: game=%s session=%s lanlan=%s err=%s",
-                archive.get("game_type"),
-                archive.get("session_id"),
-                archive.get("lanlan_name"),
-                exc,
-            )
-            return {"ok": False, "mode": "realtime", "action": "skip", "reason": "gemini_direct_response_failed"}
-        logger.info(
-            "🎮 赛后 Gemini Realtime 已直接触发: game=%s session=%s lanlan=%s bytes=%d",
-            archive.get("game_type"),
-            archive.get("session_id"),
-            archive.get("lanlan_name"),
-            len(text) + len(instruction),
-        )
-        return {
-            "ok": True,
-            "mode": "realtime",
-            "action": "direct_response",
-            "context_injected": True,
-            "nudge_scheduled": False,
-            "reason": "gemini_direct_response",
-        }
-
-    append_context = getattr(mgr, "append_context", None)
-    if not callable(append_context):
-        return {"ok": False, "mode": "realtime", "action": "skip", "reason": "context_method_unavailable"}
-    postgame_request_id = _postgame_context_request_id(archive)
-    try:
-        append_result = await append_context(
-            source="game.postgame",
-            role="system",
-            text=text,
-            audience="model",
-            timing="now",
-            lifetime="current_session",
-            request_id=postgame_request_id,
-            ordering_key=postgame_request_id,
-            metadata={
-                "game_type": archive.get("game_type"),
-                "lanlan_name": archive.get("lanlan_name"),
-                "kind": "postgame",
-            },
-        )
-    except Exception as exc:
-        logger.warning(
-            "🎮 赛后 Realtime 上下文注入失败: game=%s session=%s lanlan=%s err=%s",
-            archive.get("game_type"),
-            archive.get("session_id"),
-            archive.get("lanlan_name"),
-            exc,
-        )
-        return {"ok": False, "mode": "realtime", "action": "skip", "reason": "context_inject_failed"}
-    if not getattr(append_result, "appended", False) and not getattr(append_result, "deduped", False):
-        return {
-            "ok": False,
-            "mode": "realtime",
-            "action": "skip",
-            "reason": getattr(append_result, "reason", None) or "context_inject_failed",
-        }
-
-    logger.info(
-        "🎮 赛后 Realtime 上下文已注入: game=%s session=%s lanlan=%s bytes=%d",
-        archive.get("game_type"),
-        archive.get("session_id"),
-        archive.get("lanlan_name"),
-        len(text),
-    )
-
-    if _active_realtime_session(mgr) is not session:
-        return {
-            "ok": True,
-            "mode": "realtime",
-            "action": "context",
-            "context_injected": True,
-            "nudge_scheduled": False,
-            "nudge_reason": "realtime_session_changed",
-            "reason": "realtime_session_changed",
-            "bytes": len(text),
-        }
-
-    if getattr(append_result, "deduped", False):
-        return {
-            "ok": True,
-            "mode": "realtime",
-            "action": "context",
-            "context_injected": True,
-            "nudge_scheduled": False,
-            "reason": "context_deduped",
-        }
-
-    nudge_scheduled = False
-    nudge_reason = "disabled"
-    if options.get("trigger_voice", True):
-        trigger = getattr(mgr, "trigger_voice_proactive_nudge", None)
-        if callable(trigger):
-            asyncio.create_task(_run_postgame_realtime_nudge_task(
-                mgr,
-                dict(archive),
-                dict(options),
-                _POSTGAME_REALTIME_NUDGE_DELAYS,
-                expected_session=session,
-            ))
-            nudge_scheduled = True
-            nudge_reason = "scheduled"
-            logger.info(
-                "🎮 赛后 Realtime 主动搭话已安排: game=%s session=%s lanlan=%s delays=%s",
-                archive.get("game_type"),
-                archive.get("session_id"),
-                archive.get("lanlan_name"),
-                ",".join(f"{d:.1f}s" for d in _POSTGAME_REALTIME_NUDGE_DELAYS),
-            )
-        else:
-            nudge_reason = "trigger_unavailable"
-
-    return {
-        "ok": True,
-        "mode": "realtime",
-        "action": "nudge_scheduled" if nudge_scheduled else "context_only",
-        "context_injected": True,
-        "nudge_scheduled": nudge_scheduled,
-        "nudge_reason": nudge_reason,
-        "bytes": len(text),
-    }
-
-
-async def _deliver_postgame_text_bubble(
-    game_type: str,
-    session_id: str,
-    mgr: Any,
-    archive: dict,
-    options: dict,
-    *,
-    postgame_snapshot: Optional[dict] = None,
-) -> dict:
-    if not mgr:
-        return {"ok": False, "mode": "text", "action": "skip", "reason": "no_session_manager"}
-    if _active_realtime_session(mgr):
-        return {"ok": False, "mode": "text", "action": "skip", "reason": "active_realtime_session"}
-
-    prepare = getattr(mgr, "prepare_proactive_delivery", None)
-    finish = getattr(mgr, "finish_proactive_delivery", None)
-    if not callable(prepare) or not callable(finish):
-        return {"ok": False, "mode": "text", "action": "skip", "reason": "text_delivery_unavailable"}
-
-    try:
-        prepared = await prepare(min_idle_secs=float(options.get("min_idle_secs") or 0.0))
-    except Exception as exc:
-        logger.warning(
-            "🎮 赛后文本气泡准备失败: game=%s session=%s lanlan=%s err=%s",
-            game_type,
-            session_id,
-            archive.get("lanlan_name"),
-            exc,
-        )
-        return {"ok": False, "mode": "text", "action": "skip", "reason": "prepare_failed"}
-    if not prepared:
-        return {"ok": True, "mode": "text", "action": "pass", "reason": "condition_not_met"}
-
-    proactive_sid = getattr(mgr, "current_speech_id", None)
-    state_machine = getattr(mgr, "state", None)
-    # Why: pre-allocated out-dict captures the postgame entry/cache key
-    # AS SOON AS ``_run_game_chat`` builds the entry, so the ``finally``
-    # below can close it on EVERY termination path — including
-    # ``asyncio.CancelledError`` (which is ``BaseException`` and bypasses
-    # the structured error-result paths inside ``_run_game_chat``).
-    postgame_meta: Dict[str, Any] = {}
-    postgame_entry: Optional[dict] = None
-    postgame_cache_session_id: Optional[str] = None
-    try:
-        from main_logic.session_state import SessionEvent
-        if state_machine and hasattr(state_machine, "fire"):
-            await state_machine.fire(SessionEvent.PROACTIVE_PHASE2)
-
-        event = _build_game_postgame_event(game_type, archive, options)
-        _log_game_debug_material(
-            "postgame_text_event",
-            event,
-            game_type=game_type,
-            session_id=session_id,
-            lanlan_name=str(archive.get("lanlan_name") or ""),
-            source="game_end",
-        )
-        # Why: postgame runs AFTER ``_finalize_game_route_state`` flips the
-        # route to inactive, so the standard B1/B2 short-circuits would
-        # silently drop this designed teardown step. ``allow_postgame``
-        # opts out of those route-active gates; we own the lifecycle of
-        # any session built here and close it in the ``finally`` below.
-        llm_result = await _run_game_chat(
-            game_type, session_id, event, allow_postgame=True,
-            postgame_snapshot=postgame_snapshot,
-            postgame_meta_out=postgame_meta,
-        )
-        if isinstance(llm_result, dict):
-            postgame_entry = llm_result.get("_postgame_entry")
-            postgame_cache_session_id = llm_result.get("_postgame_cache_session_id")
-        line = str(llm_result.get("line") or "").strip()
-        if not line:
-            return {
-                "ok": True,
-                "mode": "text",
-                "action": "pass",
-                "reason": llm_result.get("error") or "empty_line",
-                "llm_source": llm_result.get("llm_source") or {},
-            }
-
-        tts_fed = False
-        feed_tts = getattr(mgr, "feed_tts_chunk", None)
-        if callable(feed_tts):
-            try:
-                await feed_tts(line, expected_speech_id=proactive_sid)
-                tts_fed = True
-            except Exception as exc:
-                logger.warning(
-                    "🎮 赛后文本气泡 TTS 投喂失败: game=%s session=%s lanlan=%s err=%s",
-                    game_type,
-                    session_id,
-                    archive.get("lanlan_name"),
-                    exc,
-                )
-
-        committed = bool(await finish(line, expected_speech_id=proactive_sid))
-        return {
-            "ok": committed,
-            "mode": "text",
-            "action": "chat" if committed else "pass",
-            "reason": "delivered" if committed else "user_took_over",
-            "line": line,
-            "turn_id": proactive_sid,
-            "tts_fed": tts_fed,
-            "llm_source": llm_result.get("llm_source") or {},
-        }
-    except Exception as exc:
-        logger.warning(
-            "🎮 赛后文本气泡投递失败: game=%s session=%s lanlan=%s err=%s",
-            game_type,
-            session_id,
-            archive.get("lanlan_name"),
-            exc,
-        )
-        return {"ok": False, "mode": "text", "action": "skip", "reason": "deliver_failed"}
-    finally:
-        try:
-            from main_logic.session_state import SessionEvent
-            if state_machine and hasattr(state_machine, "fire"):
-                await state_machine.fire(SessionEvent.PROACTIVE_DONE)
-        except Exception as exc:
-            logger.debug("🎮 赛后文本气泡状态机收尾失败: %s", exc, exc_info=True)
-        # Why: ``_run_game_chat(..., allow_postgame=True)`` builds the
-        # postgame's ``OmniOfflineClient`` at a private cache key
-        # (``::postgame::<uuid>`` suffix) so a fresh ``/route/start``
-        # reusing the user-facing ``session_id`` cannot land on the same
-        # ``_game_sessions`` slot. Identity-gating the eviction stays as
-        # defense in depth (a heartbeat sweep could theoretically pop
-        # then rebuild the postgame slot). We always close OUR captured
-        # entry's session object so the postgame client can never leak.
-        # ``OmniOfflineClient.close`` is idempotent, so a peer's prior
-        # close is safe to re-run.
-        if postgame_entry is None:
-            # Why: on ``asyncio.CancelledError`` (or any other
-            # ``BaseException``) the await above never returned, so the
-            # local ``postgame_entry``/``postgame_cache_session_id`` are
-            # still ``None``. The out-dict ``_run_game_chat`` populated
-            # mid-call still has the entry, so we fall back to it here.
-            postgame_entry = postgame_meta.get("_postgame_entry")
-            postgame_cache_session_id = postgame_meta.get("_postgame_cache_session_id")
-        if postgame_entry is not None:
-            postgame_lanlan = str(
-                postgame_entry.get("lanlan_name") or archive.get("lanlan_name") or ""
-            )
-            cache_session_id = postgame_cache_session_id or session_id
-            try:
-                key = _game_session_key(postgame_lanlan, game_type, cache_session_id)
-                cached = _game_sessions.get(key)
-                if cached is postgame_entry:
-                    _game_sessions.pop(key, None)
-                    _game_session_create_locks.pop(key, None)
-            except Exception as exc:
-                logger.debug(
-                    "🎮 赛后文本气泡 cache 清理失败: game=%s session=%s err=%s",
-                    game_type, session_id, exc, exc_info=True,
-                )
-            postgame_session = postgame_entry.get("session")
-            if postgame_session is not None:
-                try:
-                    await postgame_session.close()
-                except Exception as exc:
-                    logger.debug(
-                        "🎮 赛后文本气泡 session 清理失败: game=%s session=%s err=%s",
-                        game_type, session_id, exc, exc_info=True,
-                    )
-
-
-async def _deliver_game_postgame(
-    game_type: str,
-    session_id: str,
-    lanlan_name: str,
-    archive: dict,
-    options: dict,
-    *,
-    postgame_snapshot: Optional[dict] = None,
-) -> dict:
-    if not options.get("enabled", True):
-        return {"ok": True, "action": "skip", "reason": "disabled"}
-    mgr = get_session_manager().get(lanlan_name) if lanlan_name else None
-    mode = str(options.get("mode") or "auto").lower()
-    if mode in {"auto", "realtime"} and _active_realtime_session(mgr):
-        return await _deliver_postgame_to_realtime(mgr, archive, options)
-    if mode == "realtime":
-        return {"ok": False, "mode": "realtime", "action": "skip", "reason": "no_active_realtime_session"}
-    return await _deliver_postgame_text_bubble(
-        game_type, session_id, mgr, archive, options,
-        postgame_snapshot=postgame_snapshot,
-    )
-
-
-async def _finalize_game_route_state(
-    state: dict,
-    *,
-    reason: str,
-    close_game_session: bool = False,
-    close_debug_log: bool = True,
-) -> dict:
-    """Run the game route exit flow once, including archive submission.
-
-    Concurrent-call semantics:
-
-    - The first caller spawns ``_finalize_game_route_state_inner`` and
-      shields its task. Subsequent callers ``await asyncio.shield`` the
-      same task.
-    - ``close_game_session`` uses **OR-merge** semantics across concurrent
-      callers (B5): we stash the requested value on the state under
-      ``_exit_close_session_request``, and the inner runner reads that
-      flag (not its constructor arg) when deciding whether to close.
-      Previously the second caller's ``True`` was silently dropped while
-      the first caller's ``False`` won; the second caller then redundantly
-      invoked ``_close_and_remove_session`` outside the shield, racing
-      with the inner finalize and producing double-pop / double-close.
-    - codex P2 follow-up: a late caller arriving with
-      ``close_game_session=True`` AFTER the inner runner already passed
-      its close-site check (or finished entirely) used to lose its
-      request — the dispatcher just awaited the cached task result. We
-      now re-check ``_exit_close_session_request`` against the inner's
-      result on the existing-task path and perform the close ourselves
-      if the inner missed it. ``_close_and_remove_session`` is
-      idempotent so concurrent late callers cannot double-close.
-    """
-    if close_game_session:
-        state["_exit_close_session_request"] = True
-    elif "_exit_close_session_request" not in state:
-        state["_exit_close_session_request"] = False
-    if close_debug_log:
-        state["_exit_close_debug_log_request"] = True
-    else:
-        state["_exit_defer_debug_log_close"] = True
-        if "_exit_close_debug_log_request" not in state:
-            state["_exit_close_debug_log_request"] = False
-
-    existing_task = state.get("_exit_task")
-    if existing_task:
-        result = await asyncio.shield(existing_task)
-        if state.get("_exit_close_session_request") and not result.get("game_session_closed"):
-            closed_now = await _close_and_remove_session(
-                str(state.get("game_type") or ""),
-                str(state.get("session_id") or "default"),
-                str(state.get("lanlan_name") or ""),
-            )
-            if closed_now:
-                # Why: mutate the shared result dict so any other awaiter (or
-                # subsequent late caller) observes the close. The inner's
-                # return dict is the single source of truth handed back to
-                # every shielded await.
-                result["game_session_closed"] = True
-        if (
-            state.get("_exit_close_debug_log_request")
-            and not state.get("_exit_defer_debug_log_close")
-            and not result.get("debug_log_ended")
-        ):
-            _mark_game_session_debug_log_ended(
-                str(state.get("game_type") or ""),
-                str(state.get("session_id") or "default"),
-                lanlan_name=str(state.get("lanlan_name") or ""),
-                reason=reason,
-            )
-            result["debug_log_ended"] = True
-        return result
-
-    task = asyncio.create_task(_finalize_game_route_state_inner(state, reason=reason))
-    state["_exit_task"] = task
-    return await asyncio.shield(task)
-
-
-def _build_postgame_context_snapshot(state: dict) -> dict:
-    # Why: postgame's prompt context must be FROZEN at finalize time.
-    # Without this, ``_build_and_register_game_session`` /
-    # ``_refresh_game_session_instructions`` reverse-resolve live
-    # route_state via ``_find_game_route_state_for_session`` AFTER finalize
-    # has flipped this state inactive — and a fresh ``/route/start`` for
-    # the same ``(lanlan, game_type)`` key REPLACES the entry in
-    # ``_game_route_states``, so the lookup returns the NEW route's
-    # preGameContext / game_context. Snapshotting the two already-resolved
-    # dicts the prompt builder needs (``pre_game_context`` and
-    # ``game_context``) is the minimum-viable freeze.
-    pre_game_context = state.get("preGameContext") if isinstance(state.get("preGameContext"), dict) else None
-    return {
-        "pre_game_context": pre_game_context,
-        "game_context": _build_game_context_prompt_payload(state, include_recent=False),
-        "mode": _normalize_badminton_mode(state.get("mode")),
-    }
-
-
-async def _finalize_game_route_state_inner(
-    state: dict,
-    *,
-    reason: str,
-) -> dict:
-    state["_exit_flow_started"] = True
-    state["exit_reason"] = reason
-    state["exit_started_at"] = time.time()
-    # Capture postgame's prompt context BEFORE flipping the route inactive
-    # / before the archive resolution / before any peer ``/route/start``
-    # can replace this state in ``_game_route_states``.
-    postgame_context_snapshot = _build_postgame_context_snapshot(state)
-    state["game_route_active"] = False
-    state["game_external_voice_route_active"] = False
-    state["game_external_text_route_active"] = False
-    state["heartbeat_enabled"] = False
-    lanlan_name = str(state.get("lanlan_name") or "")
-    mgr = get_session_manager().get(lanlan_name) if lanlan_name else None
-    # 推 closed 事件让前端还原 chat.html 折叠态 + 显回 pet 容器。所有 finalize
-    # 路径（/route/end / heartbeat sweep / supersede）都走本 inner，与 active
-    # flag 翻 false 同源，不会出现"已结束但 UI 仍锁着收缩态"的孤岛。
-    await _push_game_window_state_change(
-        mgr,
-        action="closed",
-        lanlan_name=lanlan_name,
-        game_type=str(state.get("game_type") or ""),
-        session_id=str(state.get("session_id") or ""),
-    )
-    # Release the SessionManager-level takeover so ordinary chat handlers come
-    # back online; chat LLM may produce auto-replies again, but the player has
-    # exited the game so that's the desired behavior.
-    if mgr is not None:
-        mgr._takeover_active = False
-        mgr._takeover_input_dispatcher = None
-    realtime_restore = {"attempted": False, "ok": True, "reason": "takeover_released"}
-    state["realtime_restore"] = realtime_restore
-    if mgr and hasattr(mgr, "send_status"):
-        try:
-            await mgr.send_status(json.dumps({
-                "code": "GAME_ROUTE_ENDED",
-                "details": {
-                    "game_type": str(state.get("game_type") or ""),
-                    "session_id": str(state.get("session_id") or ""),
-                    "lanlan_name": lanlan_name,
-                    "reason": reason,
-                    "before_game_external_mode": state.get("before_game_external_mode"),
-                    "before_game_external_active": bool(state.get("before_game_external_active")),
-                    "should_resume_external_on_exit": bool(state.get("should_resume_external_on_exit")),
-                    "realtime_restore": realtime_restore,
-                },
-            }))
-        except Exception as exc:
-            logger.warning("⚠️ 游戏路由退出状态通知失败: %s", exc)
-
-    skip_memory_reason = _game_archive_memory_skip_reason(state, reason)
-    if skip_memory_reason == "game_memory_archive_disabled":
-        await _cancel_game_context_organizer_before_disabled_archive(state)
-    else:
-        await _settle_game_context_organizer_before_archive(state)
-
-    archive = state.get("archive") if isinstance(state.get("archive"), dict) else None
-    if archive is None:
-        archive = _build_game_archive(state)
-    archive["exit_reason"] = reason
-    state["archive"] = archive
-
-    memory_result = state.get("archive_memory_result")
-    if not isinstance(memory_result, dict):
-        if skip_memory_reason:
-            archive["memory_skipped"] = True
-            archive["memory_skip_reason"] = skip_memory_reason
-            memory_result = _build_game_archive_memory_skipped_result(skip_memory_reason)
-        else:
-            memory_result = await _submit_game_archive_to_memory(archive)
-        state["archive_memory_result"] = memory_result
-
-    # B5: OR-merge close decision across concurrent callers (see note in
-    # ``_finalize_game_route_state``). Re-read the flag *after* awaiting
-    # the archive work so a second caller arriving mid-finalize with
-    # ``close_game_session=True`` still wins.
-    session_closed = False
-    if state.get("_exit_close_session_request"):
-        session_closed = await _close_and_remove_session(
-            str(state.get("game_type") or ""),
-            str(state.get("session_id") or "default"),
-            str(state.get("lanlan_name") or ""),
-        )
-    debug_log_ended = False
-    if state.get("_exit_close_debug_log_request") and not state.get("_exit_defer_debug_log_close"):
-        _mark_game_session_debug_log_ended(
-            str(state.get("game_type") or ""),
-            str(state.get("session_id") or "default"),
-            lanlan_name=str(state.get("lanlan_name") or ""),
-            reason=reason,
-        )
-        debug_log_ended = True
-
-    return {
-        "archive": archive,
-        "archive_memory": memory_result,
-        "game_session_closed": session_closed,
-        "debug_log_ended": debug_log_ended,
-        "exit_reason": reason,
-        "realtime_restore": realtime_restore,
-        "postgame_context_snapshot": postgame_context_snapshot,
-    }
-
-
-async def _get_or_create_session(
-    game_type: str,
-    session_id: str,
-    lanlan_name: str = "",
-    *,
-    postgame_snapshot: Optional[dict] = None,
-) -> dict:
-    """Get or create a game session.
-
-    B6: serialize the cache-miss → ctor → connect → cache-insert sequence
-    under a per-key ``asyncio.Lock`` so two concurrent ``_run_game_chat``
-    calls for the same ``(lanlan, game_type, session_id)`` cannot both
-    build a fresh ``OmniOfflineClient`` and overwrite each other in
-    ``_game_sessions``, leaking the loser's connection.
-
-    CodeRabbit follow-up: ``lanlan_name`` may be empty on entry
-    (caller-supplied) but canonicalizes to ``char_info["lanlan_name"]``.
-    Resolve the canonical key BEFORE acquiring the create lock so we
-    only ever take one lock — under the canonical key. The previous
-    "lock under raw key, then re-lock under canonical key" shape left
-    an orphan ``_game_session_create_locks[raw_key]`` entry whenever
-    the canonical resolution changed the key, because
-    ``_close_and_remove_session`` only evicts the lock keyed by the
-    session's actual storage key (the canonical one).
-
-    The fast-path cache check still uses the raw key so a hit on the
-    pre-canonicalization shape (rare; only happens if a session was
-    cached under an empty lanlan_name) short-circuits without paying
-    the ``_get_character_info`` lookup.
-    """
-    key = _game_session_key(lanlan_name, game_type, session_id)
-
-    if key in _game_sessions:
-        entry = _game_sessions[key]
-        entry['last_activity'] = time.time()
-        return entry
-
-    char_info = _get_character_info(lanlan_name)
-    canonical_lanlan = str(char_info.get("lanlan_name") or lanlan_name or "").strip()
-    canonical_key = _game_session_key(canonical_lanlan, game_type, session_id)
-
-    # Fast path: canonical_key may already be cached (e.g. another caller
-    # passed the canonical lanlan_name and built it).
-    if canonical_key in _game_sessions:
-        entry = _game_sessions[canonical_key]
-        entry['last_activity'] = time.time()
-        return entry
-
-    create_lock = _get_session_create_lock(canonical_key)
-    async with create_lock:
-        if canonical_key in _game_sessions:
-            entry = _game_sessions[canonical_key]
-            entry['last_activity'] = time.time()
-            return entry
-        try:
-            return await _build_and_register_game_session(
-                canonical_key, game_type, session_id, char_info,
-                postgame_snapshot=postgame_snapshot,
-            )
-        except BaseException:
-            # codex P2 (PR #1127 r3182157092): if the build raises,
-            # nothing inserts into ``_game_sessions`` for this key, so
-            # ``_close_and_remove_session`` will never run for it and
-            # the per-key create lock would leak forever. Evict it
-            # here.
-            #
-            # Why conditional pop on ``_waiters``: if a peer task is
-            # already awaiting THIS lock (concurrent miss for the same
-            # canonical_key), unconditionally popping would let a new
-            # arrival call ``_get_session_create_lock`` and receive a
-            # FRESH lock object — distinct from the one the peer is
-            # awaiting — defeating the build serialization. Leaving
-            # the lock in place when there are waiters keeps them on
-            # the same Lock instance; the next successful build path
-            # registers an entry and ``_close_and_remove_session``
-            # will pop normally; another failed build will hit this
-            # branch again and eventually find ``_waiters`` empty.
-            # ``_waiters`` is CPython-private but stable across all
-            # supported Python versions.
-            waiters = getattr(create_lock, "_waiters", None)
-            if not waiters:
-                _game_session_create_locks.pop(canonical_key, None)
-            raise
-
-
-async def _build_and_register_game_session(
-    key: str,
-    game_type: str,
-    session_id: str,
-    char_info: dict,
-    *,
-    postgame_snapshot: Optional[dict] = None,
-) -> dict:
-    """Build a fresh game session entry; caller must already hold the
-    per-key creation lock (see ``_get_or_create_session``).
-
-    ``postgame_snapshot`` (when set) is the authoritative prompt-context
-    source for postgame builds — see ``_build_postgame_context_snapshot``.
-    Without it, a postgame build that races a fresh ``/route/start`` for
-    the same ``(lanlan, game_type, session_id)`` would reverse-resolve
-    live route_state and pick up the NEW route's preGameContext /
-    game_context.
-    """
-    from main_logic.omni_offline_client import OmniOfflineClient
-    from utils.token_tracker import set_call_type
-
-    lanlan_name = str(char_info.get("lanlan_name") or "").strip()
-    reply_chunks: list[str] = []
-
-    async def on_text_delta(text: str, is_first: bool, **_kwargs):
-        # **_kwargs 吞掉 ui_enabled / tts_enabled（OmniOfflineClient summary 路径会传，
-        # 但 game 短台词跑的是非 summary 模式，理论上不会发 UI/TTS 分流。保留 kwargs
-        # forward-compat 防签名漂移触发 TypeError。）
-        reply_chunks.append(text)
-
-    set_call_type("game_chat")
-
-    session = OmniOfflineClient(
-        base_url=char_info['base_url'],
-        api_key=char_info['api_key'],
-        model=char_info['model'],
-        on_text_delta=on_text_delta,
-        max_response_length=100,  # 游戏台词要短
-        lanlan_name=char_info['lanlan_name'],
-        master_name=char_info['master_name'],
-    )
-
-    if postgame_snapshot is not None:
-        pre_game_context = postgame_snapshot.get("pre_game_context")
-        game_context = postgame_snapshot.get("game_context")
-        route_mode = _normalize_badminton_mode(postgame_snapshot.get("mode"))
-    else:
-        route_state = _find_game_route_state_for_session(game_type, _route_session_id(session_id), lanlan_name)
-        pre_game_context = route_state.get("preGameContext") if isinstance(route_state, dict) else None
-        game_context = _build_game_context_prompt_payload(route_state, include_recent=False)
-        route_mode = _normalize_badminton_mode(route_state.get("mode") if isinstance(route_state, dict) else "")
-    prompt_args = (
-        game_type,
-        char_info['lanlan_name'],
-        char_info['lanlan_prompt'],
-        pre_game_context if isinstance(pre_game_context, dict) else None,
-        game_context if isinstance(game_context, dict) else None,
-        char_info.get("user_language"),
-    )
-    if _is_badminton_game_type(game_type):
-        system_prompt = _build_game_prompt(*prompt_args, mode=route_mode)
-    else:
-        system_prompt = _build_game_prompt(*prompt_args)
-    try:
-        await session.connect(instructions=system_prompt)
-    except asyncio.CancelledError:
-        # Why: CancelledError doesn't inherit from Exception in Python
-        # 3.8+; without this branch a cancelled connect leaks the half-
-        # open client.
-        try:
-            await session.close()
-        except Exception:
-            # Why: cleanup must remain idempotent on the cancellation
-            # path — a close() failure here would mask the original
-            # CancelledError that we re-raise below.
-            pass
-        raise
-    except Exception:
-        # Connect failed — ensure we don't leak a half-open client. close
-        # is idempotent / tolerant of "never connected".
-        try:
-            await session.close()
-        except Exception:
-            # Why: cleanup must not raise from the exception path —
-            # a close() failure would mask the original connect error
-            # that we re-raise below.
-            pass
-        raise
-
-    entry = {
-        'session': session,
-        'reply_chunks': reply_chunks,
-        'lanlan_name': char_info['lanlan_name'],
-        'lanlan_prompt': char_info.get('lanlan_prompt') or '',
-        'user_language': char_info.get('user_language'),
-        'source': _infer_service_source(
-            char_info.get('base_url', ''),
-            char_info.get('model', ''),
-            char_info.get('api_type', ''),
-        ),
-        'last_activity': time.time(),
-        'lock': asyncio.Lock(),
-        'instructions': system_prompt,
-        'mode': route_mode,
-    }
-    _game_sessions[key] = entry
-
-    logger.info(
-        "🎮 创建游戏LLM会话: 游戏=%s 会话=%s 角色=%s 模型=%s 人格提示长度=%d字",
-        game_type,
-        session_id,
-        char_info['lanlan_name'],
-        char_info['model'],
-        len(char_info.get('lanlan_prompt') or ''),
-    )
-    return entry
-
-
-async def _refresh_game_session_instructions(
-    entry: dict,
-    game_type: str,
-    session_id: str,
-    lanlan_name: str = "",
-    *,
-    postgame_snapshot: Optional[dict] = None,
-) -> None:
-    session = entry.get("session") if isinstance(entry, dict) else None
-    update = getattr(session, "update_session", None)
-    if not callable(update):
-        return
-
-    lanlan_name = str(lanlan_name or entry.get("lanlan_name") or "").strip()
-    char_info = _get_character_info(lanlan_name)
-    entry["user_language"] = char_info.get("user_language")
-    if postgame_snapshot is not None:
-        pre_game_context = postgame_snapshot.get("pre_game_context")
-        game_context = postgame_snapshot.get("game_context")
-        route_mode = _normalize_badminton_mode(postgame_snapshot.get("mode"))
-    else:
-        route_state = _find_game_route_state_for_session(game_type, _route_session_id(session_id), char_info["lanlan_name"])
-        pre_game_context = route_state.get("preGameContext") if isinstance(route_state, dict) else None
-        game_context = _build_game_context_prompt_payload(route_state, include_recent=False)
-        route_mode = _normalize_badminton_mode(route_state.get("mode") if isinstance(route_state, dict) else "")
-    prompt_args = (
-        game_type,
-        char_info["lanlan_name"],
-        char_info["lanlan_prompt"],
-        pre_game_context if isinstance(pre_game_context, dict) else None,
-        game_context if isinstance(game_context, dict) else None,
-        char_info.get("user_language"),
-    )
-    if _is_badminton_game_type(game_type):
-        instructions = _build_game_prompt(*prompt_args, mode=route_mode)
-    else:
-        instructions = _build_game_prompt(*prompt_args)
-    if entry.get("instructions") == instructions:
-        entry["mode"] = route_mode
-        return
-    await update({"instructions": instructions})
-    entry["instructions"] = instructions
-    entry["mode"] = route_mode
 
 
 def _parse_control_instructions(reply: str, game_type: str = "soccer") -> Dict[str, Any]:
@@ -2436,13 +517,6 @@ def _parse_control_instructions(reply: str, game_type: str = "soccer") -> Dict[s
     }
 
 
-def _game_session_key(lanlan_name: str, game_type: str, session_id: str) -> str:
-    lanlan = str(lanlan_name or "").strip()
-    if lanlan:
-        return f"{lanlan}:{game_type}:{session_id}"
-    return f"{game_type}:{session_id}"
-
-
 def _strip_ssml_like_tags(text: str) -> str:
     """Remove known SSML tags before handing text to TTS."""
     line = str(text or "")
@@ -2464,134 +538,6 @@ def _check_badminton_chat_rate(lanlan_name: str, session_id: str) -> bool:
     _badminton_chat_rate_windows.move_to_end(key)
     while len(_badminton_chat_rate_windows) > 128:
         _badminton_chat_rate_windows.popitem(last=False)
-    return True
-
-
-_POSTGAME_SESSION_MARKER = "::postgame::"
-
-
-# Why: ``_make_postgame_session_id`` produces ``<session_id>::postgame::<uuid4.hex>``
-# where ``uuid4().hex`` is exactly 32 lowercase hex chars. ``_route_session_id``
-# strips ONLY this exact synthetic suffix shape so a legitimate client-supplied
-# session_id that happens to contain ``::postgame::`` is left untouched.
-_POSTGAME_UUID_TAIL_RE = re.compile(r"[0-9a-f]{32}\Z")
-
-
-def _make_postgame_session_id(session_id: str) -> str:
-    # Why: postgame's freshly-built session lives at a private cache key
-    # so a racing ``/route/start`` reusing the same user-facing session_id
-    # cannot land on the same ``_game_sessions`` slot. Without this, a
-    # peer's first ``/game_chat`` would be handed back postgame's cached
-    # entry by ``_get_or_create_session``, and postgame's ``finally``
-    # close (still identity-matching since the cache wasn't replaced)
-    # would tear down the active route's session mid-turn.
-    return f"{str(session_id or '')}{_POSTGAME_SESSION_MARKER}{uuid.uuid4().hex}"
-
-
-def _route_session_id(session_id: str) -> str:
-    # Why: this helper is now defensive. The critical postgame paths
-    # (``_build_and_register_game_session`` / ``_refresh_game_session_instructions``)
-    # use a frozen ``postgame_snapshot`` instead of reverse-resolving live
-    # route_state, so the marker no longer needs to round-trip through
-    # ``_find_game_route_state_for_session`` for prompt context. We still
-    # tolerate the synthetic shape elsewhere by stripping ONLY the exact
-    # suffix produced by ``_make_postgame_session_id`` (marker + 32-char
-    # uuid4 hex tail at end of string). A legitimate client session_id
-    # that happens to contain ``::postgame::`` mid-string — or with a
-    # non-uuid tail — is returned unchanged.
-    raw = str(session_id or "")
-    idx = raw.rfind(_POSTGAME_SESSION_MARKER)
-    if idx == -1:
-        return raw
-    tail = raw[idx + len(_POSTGAME_SESSION_MARKER):]
-    if not _POSTGAME_UUID_TAIL_RE.fullmatch(tail):
-        return raw
-    return raw[:idx]
-
-
-def _parse_game_session_key(key: str) -> tuple[str, str, str]:
-    parts = str(key or "").split(":", 2)
-    if len(parts) == 3:
-        return parts[0], parts[1], parts[2]
-    game_type, _, session_id = str(key or "").partition(":")
-    return "", game_type, session_id
-
-
-async def _close_and_remove_session(
-    game_type: str,
-    session_id: str,
-    lanlan_name: str = "",
-) -> bool:
-    """Close and remove the specified game session.
-
-    B1: serialize against in-flight ``_run_game_chat`` work for the same
-    entry by acquiring ``entry['lock']`` before popping + closing. Without
-    this, a concurrent close (from ``/route/start`` finalize, heartbeat
-    sweep, or ``/route/end``) would yank the session out of the cache and
-    close it while a chat call still held a reference and was mid
-    ``stream_text``, producing reads against a closed client.
-
-    The entry-level lock keeps the wait bounded — chat work is capped by
-    the 15s ``stream_text`` timeout. New chats arriving after we set the
-    route's ``_exit_flow_started`` flag short-circuit before they ever
-    touch the entry lock.
-
-    codex P1 (PR #1127 r3182582714): identity-gate the cache eviction.
-    Two close callers can read the same ``entry`` then queue on its
-    lock; while they wait, a peer may pop the cache, a fresh
-    ``/route/start`` may build ``entry_NEW`` under the same key, and we
-    would otherwise wake up and pop ``entry_NEW`` — closing a live
-    session a different route owns. Mirrors the postgame ownership gate
-    in ``_deliver_postgame_text_bubble``'s finally. We always close OUR
-    captured ``entry``'s session (it's ours since the top of this
-    function) but only touch the cache / create-lock dicts if they
-    still point at us.
-    """
-    keys = []
-    if lanlan_name:
-        keys.append(_game_session_key(lanlan_name, game_type, session_id))
-    keys.append(_game_session_key("", game_type, session_id))
-
-    # First locate the entry (without popping) to grab its lock, then pop
-    # under the lock. Two close callers racing here both serialize on the
-    # same lock and only one observes a non-None entry after pop.
-    key = ""
-    entry = None
-    for candidate in keys:
-        key = candidate
-        entry = _game_sessions.get(candidate)
-        if entry:
-            break
-    if not entry:
-        return False
-
-    entry_lock = entry.get('lock')
-    if isinstance(entry_lock, asyncio.Lock):
-        async with entry_lock:
-            cache_owned = _game_sessions.get(key) is entry
-            if cache_owned:
-                _game_sessions.pop(key, None)
-                _game_session_create_locks.pop(key, None)
-    else:
-        cache_owned = _game_sessions.get(key) is entry
-        if cache_owned:
-            _game_sessions.pop(key, None)
-            _game_session_create_locks.pop(key, None)
-
-    # Why: ``entry`` was captured at the top of this function; we own its
-    # lifecycle even if a peer closer rotated the cache to ``entry_NEW``
-    # while we waited on the lock. Always close OUR session so the
-    # client cannot leak. ``OmniOfflineClient.close`` is idempotent
-    # (omni_offline_client.py:1815-1835), so any peer's prior close on
-    # the same object is safe to re-run.
-    session = entry.get('session')
-    if session:
-        try:
-            await session.close()
-        except Exception as e:
-            logger.debug("🎮 关闭游戏 session 失败: key=%s err=%s", key, e, exc_info=True)
-
-    logger.info("🎮 结束游戏 session: %s cache_owned=%s", key, cache_owned)
     return True
